use correct type for arm_tape_halt_index

+++ sim.h
@@ -322,7 +322,7 @@ struct solution {
     int64_t *arm_tape_start_cycle;
     // the instruction index after which to halt execution of each tape.  set to
     // SIZE_MAX if the arm never halts.
-    int64_t *arm_tape_halt_index;
+    size_t *arm_tape_halt_index;
     size_t number_of_arms;
 
     // the maximum absolute value of all arm rotation amounts over all cycles.
